[Chore] #108 Auth 관련 파일명 수정
